Move legacy rewriters to own package
