+++ okapi-ir/src/main/scala/org/opencypher/okapi/ir/impl/parse/rewriter/legacy/normalizeWithClauses.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2002-2018 Neo4j Sweden AB (http://neo4j.com)
+ * Copyright (c) 2016-2018 "Neo4j Sweden, AB" [https://neo4j.com]
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,10 +12,21 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ * Attribution Notice under the terms of the Apache License 2.0
+ *
+ * This work was created by the collective efforts of the openCypher community.
+ * Without limiting the terms of Section 6, any Derivative Work that is not
+ * approved by the public consensus process of the openCypher Implementers Group
+ * should not be described as “Cypher” (and Cypher® is a registered trademark of
+ * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
+ * proposals for change that have been documented or implemented should only be
+ * described as "implementation extensions to Cypher" or as "proposed changes to
+ * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.v9_0.rewriting.rewriters
+package org.opencypher.okapi.ir.impl.parse.rewriter.legacy
 
-import org.opencypher.v9_0.ast.{Where, _}
+import org.opencypher.v9_0.ast._
 import org.opencypher.v9_0.expressions.{Expression, LogicalVariable}
 import org.opencypher.v9_0.util._
 
@@ -41,7 +52,7 @@ import org.opencypher.v9_0.util._
   * WITH n.prop AS prop ORDER BY prop DESC
   * RETURN prop
   */
-case class legacyNormalizeWithClauses(mkException: (String, InputPosition) => CypherException) extends Rewriter {
+case class normalizeWithClauses(mkException: (String, InputPosition) => CypherException) extends Rewriter {
 
   def apply(that: AnyRef): AnyRef = instance.apply(that)
